refactor(git-review): nest context helpers for dir-density cap
Move panel context synthesis and related unit tests under feature
subfolders so review directories stay within the hard density limit.
